[order] Add type, model, location when generating a new item
git-svn-id: https://forge.glpi-project.org/svn/order/trunk@128 349b9182-4a13-0410-896f-e5e9767dd1b3

+++ inc/plugin_order.dropdown.function.php
@@ -139,13 +139,15 @@ function plugin_order_dropdownAllItemsByType($name, $type, $entity=0,$item_type=
 	return dropdownArrayValues($name, $items, 0);
 }
 
-function plugin_order_dropdownReceptionActions($type,$referenceID) {
+function plugin_order_dropdownReceptionActions($type,$referenceID,$orderID) {
 	global $LANG, $CFG_GLPI;
 	$rand = mt_rand();
 	echo "<td width='5%'>";
 	echo "<select name='receptionActions$rand' id='receptionActions$rand'>";
 	echo "<option value='0' selected>-----</option>";
-	echo "<option value='reception'>" . $LANG['plugin_order']['delivery'][2] . "</option>";
+	if (!plugin_order_allItemsAlreadyDelivered($orderID, $referenceID))
+		echo "<option value='reception'>" . $LANG['plugin_order']['delivery'][2] . "</option>";
+		
 	if ($type != CONSUMABLE_ITEM_TYPE && $type != CARTRIDGE_ITEM_TYPE)
 		echo "<option value='generation'>" . $LANG['plugin_order']['delivery'][3] . "</option>";
 	echo "<option value='createLink'>" . $LANG['plugin_order']['delivery'][11] . "</option>";
